+++ src/lte/model/lte-as-sap.h
@@ -56,7 +56,7 @@ public:
    *
    * \param dlEarfcn the downlink carrier frequency (EARFCN)
    */
-  virtual void StartCellSelection (uint16_t dlEarfcn) = 0;
+  virtual void StartCellSelection (uint32_t dlEarfcn) = 0;
 
   /** 
    * \brief Force the RRC entity to stay camped on a certain eNodeB.
@@ -64,7 +64,7 @@ public:
    * \param cellId the cell ID identifying the eNodeB
    * \param dlEarfcn the downlink carrier frequency (EARFCN)
    */
-  virtual void ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn) = 0;
+  virtual void ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn) = 0;
 
   /**
    * \brief Tell the RRC entity to enter Connected mode.
@@ -151,8 +151,8 @@ public:
 
   // inherited from LteAsSapProvider
   virtual void SetCsgWhiteList (uint32_t csgId);
-  virtual void StartCellSelection (uint16_t dlEarfcn);
-  virtual void ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn);
+  virtual void StartCellSelection (uint32_t dlEarfcn);
+  virtual void ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn);
   virtual void Connect (void);
   virtual void SendData (Ptr<Packet> packet, uint8_t bid);
   virtual void Disconnect ();
@@ -182,14 +182,14 @@ MemberLteAsSapProvider<C>::SetCsgWhiteList (uint32_t csgId)
 
 template <class C>
 void
-MemberLteAsSapProvider<C>::StartCellSelection (uint16_t dlEarfcn)
+MemberLteAsSapProvider<C>::StartCellSelection (uint32_t dlEarfcn)
 {
   m_owner->DoStartCellSelection (dlEarfcn);
 }
 
 template <class C>
 void
-MemberLteAsSapProvider<C>::ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn)
+MemberLteAsSapProvider<C>::ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn)
 {
   m_owner->DoForceCampedOnEnb (cellId, dlEarfcn);
 }
