+++ src/lte/model/lte-ue-phy.h
@@ -310,11 +310,11 @@ private:
 
   // UE CPHY SAP methods
   void DoReset ();
-  void DoStartCellSearch (uint16_t dlEarfcn);
+  void DoStartCellSearch (uint32_t dlEarfcn);
   void DoSynchronizeWithEnb (uint16_t cellId);
-  void DoSynchronizeWithEnb (uint16_t cellId, uint16_t dlEarfcn);
+  void DoSynchronizeWithEnb (uint16_t cellId, uint32_t dlEarfcn);
   void DoSetDlBandwidth (uint8_t ulBandwidth);
-  void DoConfigureUplink (uint16_t ulEarfcn, uint8_t ulBandwidth);
+  void DoConfigureUplink (uint32_t ulEarfcn, uint8_t ulBandwidth);
   void DoConfigureReferenceSignalPower (int8_t referenceSignalPower);
   void DoSetRnti (uint16_t rnti);
   void DoSetTransmissionMode (uint8_t txMode);
